Time: 23 ms (44.57%), Space: 54.8 MB (14.13%) - LeetHub

+++ 1987-number-of-unique-good-subsequences/1987-number-of-unique-good-subsequences.java
@@ -1,12 +1,12 @@
 class Solution {
     public int numberOfUniqueGoodSubsequences(String binary) {
-        long[] count=new long[2];
+        int[] count=new int[2];
         
         int mod=(int)(Math.pow(10,9)+7);
         
         for(char ch:binary.toCharArray()){
             if(ch=='0'){
-                count[0] = ((count[0]>0?count[0]-1:0) +count[1] + 1)%mod; 
+                count[0] = ((count[0]>0?count[0]-1:0) + count[1] + 1)%mod; 
             }else{
                 count[1] = ((count[0]>0?count[0]-1:0) + count[1] + 1)%mod;
             }
